# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeMarket API",
    "version": "1.3.0",
    "description": "WeMarket QR Menu & Small Business Platform API"
  },
  "servers": [
    {
      "url": "https://wemarket.onrender.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "orderCapability": {
        "type": "apiKey",
        "in": "header",
        "name": "x-order-capability"
      }
    }
  },
...
  "paths": {
...
    "/api/categories/store/{storeId}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "매장별 카테고리 조회",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장 카테고리 목록 반환"
          }
        }
      }
    },
...
    "/api/eco-badge/store/{storeId}": {
      "get": {
        "tags": [
          "Eco Badge"
        ],
        "summary": "매장 친환경 뱃지 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "친환경 뱃지 정보"
          }
        }
      }
    },
...
    "/api/food-trucks/stores/{storeId}": {
      "get": {
        "tags": [
          "FoodTrucks"
        ],
        "summary": "푸드트럭 개별 실시간 상태 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "푸드트럭 상태 반환"
          }
        }
      }
    },
...
    "/api/inventory/store/{storeId}": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "매장 재고 현황 목록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "재고 목록"
          }
        }
      }
    },
...
    "/api/legal/stores/{storeId}": {
      "get": {
        "tags": [
          "Legal"
        ],
        "summary": "전자상거래법 필수 표시 사항 조회",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "필수 표시 사항"
          }
        }
      }
    },
...
    "/api/legal/admin/stores/{storeId}": {
      "get": {
        "tags": [
          "Legal"
        ],
        "summary": "전체 법적 정보 조회 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "법적 정보 전체"
          }
        }
      },
      "put": {
        "tags": [
          "Legal"
        ],
        "summary": "법적 정보 수정 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 완료"
          }
        }
      }
    },
...
    "/api/naver-place/store/{storeId}": {
      "get": {
        "tags": [
          "NaverPlace"
        ],
        "summary": "매장의 네이버 플레이스 정보 조회",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "네이버 플레이스 정보 (매장명, 별점, 리뷰URL 등)"
          },
          "404": {
            "description": "매장 미발견"
          },
          "503": {
            "description": "NAVER_CLIENT_SECRET 미설정"
          }
        }
      }
    },
...
    "/api/option-templates/store/{storeId}": {
      "get": {
        "tags": [
          "OptionTemplates"
        ],
        "summary": "매장별 옵션 템플릿 목록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "옵션 템플릿 목록"
          }
        }
      }
    },
...
    "/api/orders/store/{storeId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "매장별 주문 목록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 목록 (페이지네이션)"
          }
        }
      }
    },
...
    "/api/plan-requests/store/{storeId}": {
      "get": {
        "tags": [
          "PlanRequests"
        ],
        "summary": "내 매장의 신청 내역 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장별 신청 목록"
          }
        }
      }
    },
...
    "/api/print-jobs/store/{storeId}": {
      "get": {
        "tags": [
          "PrintJobs"
        ],
        "deprecated": true,
        "summary": "(하위호환) 매장별 미처리 작업 조회 — storeId는 무시되고 API 키 매장이 사용된다",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "미처리 작업 목록"
          },
          "403": {
            "description": "API 키 매장과 요청 매장 불일치"
          }
        }
      }
    },
...
    "/api/products/store/{storeId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "매장별 상품 목록 조회 (메뉴판)",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상품 목록 (60초 캐시)"
          }
        }
      }
    },
...
    "/api/reports/store/{storeId}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "매장 매출 리포트 PDF 생성 및 다운로드",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF 파일 다운로드",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reservations/store/{storeId}": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "summary": "매장 예약 목록 조회 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "cancelled",
                "completed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예약 목록"
          }
        }
      }
    },
...
    "/api/reviews/store/{storeId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "매장별 리뷰 목록 조회 (최신순)",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "리뷰 목록"
          }
        }
      }
    },
...
    "/api/staff/store/{storeId}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "summary": "매장 직원 목록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "직원 목록"
          }
        }
      }
    },
...
    "/api/staff-requests/store/{storeId}": {
      "get": {
        "tags": [
          "StaffRequests"
        ],
        "summary": "매장별 신청 내역 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장별 신청 목록"
          }
        }
      }
    },
...
    "/api/stores": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "전체 매장 목록 조회",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장 목록"
          }
        }
      },
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "매장 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "매장 생성 완료"
          }
        }
      }
    },
    "/api/stores/search": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "매장 검색",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "검색어 (이름, 주소)"
          }
        ],
        "responses": {
          "200": {
            "description": "검색 결과"
          }
        }
      }
    },
...
    "/api/stores/favorites/{phone}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "관심 매장 목록 조회",
        "parameters": [
          {
            "in": "path",
            "name": "phone",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "관심 매장 목록"
          }
        }
      }
    },
    "/api/stores/favorites": {
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "관심 매장 추가",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone",
                  "store_id"
                ],
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "store_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "추가 완료"
          }
        }
      },
      "delete": {
        "tags": [
          "Stores"
        ],
        "summary": "관심 매장 삭제",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone",
                  "store_id"
                ],
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "store_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "삭제 완료"
          }
        }
      }
    },
    "/api/stores/my": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "내 매장 목록 조회 (인증 필요)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "관리하는 매장 목록"
          }
        }
      }
    },
    "/api/stores/{id}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "매장 상세 조회",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장 상세 정보"
          }
        }
      },
      "put": {
        "tags": [
          "Stores"
        ],
        "summary": "매장 정보 수정 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 완료"
          }
        }
      },
      "delete": {
        "tags": [
          "Stores"
        ],
        "summary": "매장 삭제 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 완료"
          }
        }
      }
    },
    "/api/stores/{id}/business": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "사업자 정보 조회 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사업자 정보"
          }
        }
      },
      "put": {
        "tags": [
          "Stores"
        ],
        "summary": "사업자 정보 수정 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessInfoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 완료"
          }
        }
      }
    },
    "/api/stores/{id}/account": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "정산 계좌 조회 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정산 계좌 정보"
          }
        }
      },
      "put": {
        "tags": [
          "Stores"
        ],
        "summary": "정산 계좌 수정 (관리자)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 완료"
          }
        }
      }
    },
...
    "/api/v1/store": {
      "get": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "매장 정보 조회",
        "security": [
          {
            "apiKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "매장 정보"
          }
        }
      }
    },
...
  }
}